+++ server/BraveAlerterConfigurator.js
@@ -109,7 +109,7 @@ class BraveAlerterConfigurator {
           }
 
           if (alertSession.incidentCategoryKey) {
-            session.incidentCategory = session.device.client.incidentCategories[parseInt(alertSession.incidentCategoryKey, 10) - 1]
+            session.incidentCategory = session.device.client.incidentCategories[parseInt(alertSession.incidentCategoryKey, 10)]
           }
 
           if (alertSession.alertState === CHATBOT_STATE.RESET) {
@@ -229,7 +229,7 @@ class BraveAlerterConfigurator {
   createIncidentCategoryKeys(incidentCategories) {
     // Incident categories in Sensors are always 1-indexed
     const incidentCategoryKeys = []
-    for (let i = 1; i <= incidentCategories.length; i += 1) {
+    for (let i = 0; i < incidentCategories.length; i += 1) {
       incidentCategoryKeys.push(i.toString())
     }
 
@@ -239,7 +239,7 @@ class BraveAlerterConfigurator {
   createResponseStringFromIncidentCategories(categories, language) {
     function reducer(accumulator, currentValue, currentIndex) {
       // Incident categories in Sensors are always 1-indexed
-      return `${accumulator}${currentIndex + 1} - ${t(currentValue, { lng: language })}\n`
+      return `${accumulator}${currentIndex} - ${t(currentValue, { lng: language })}\n`
     }
     return categories.reduce(reducer, '')
   }